Move sources to src/ and devices/, add CMake libraries
