+++ src/devices/lis3mdl.cc
