+++ src/devices/lis3mdl.h
